+++ client/src/components/auth/Login.js
@@ -3,7 +3,7 @@ import { Link, Redirect } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faUser } from '@fortawesome/free-solid-svg-icons';
 
-import { UserContext } from '../../UserContext';
+import { UserContext } from '../../contexts/UserContext';
 
 const Login = () => {
     const [formData, setFormData] = useState({