+++ client/src/contexts/UserContext.js
@@ -2,7 +2,7 @@ import React, { useState, useContext } from 'react';
 import axios from 'axios';
 
 import { AlertContext } from './AlertContext';
-import setAuthToken from './utils/setAuthToken';
+import setAuthToken from '../utils/setAuthToken';
 
 const UserContext = React.createContext();
 
@@ -83,6 +83,19 @@ function UserProvider({ children }) {
         }
     }
 
+    async function deleteAccount() {
+        if (window.confirm('Are you sure? This can NOT be undone!')) {
+            try {
+                await axios.delete(`api/profile/`);
+                setAlert('Account was deleted.', 'danger');
+                authError();
+            } catch (err) {
+                const errors = err.response.data.errors;
+                authError(errors);
+            }
+        }
+    }
+
     return (
         <UserContext.Provider
             value={{
@@ -90,7 +103,8 @@ function UserProvider({ children }) {
                 register,
                 authError,
                 loadUser,
-                login
+                login,
+                deleteAccount
             }}
         >
             {children}